requirements.txt
streamlit
openai-whisper
sentence-transformers
language-tool-python
vaderSentiment
google-generativeai
plotly
pydub
ffmpeg-python
watchdog
